data_sources/semantic_scholar/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.7
    # via requests
click==8.0.3
    # via typer
idna==3.3
    # via requests
kafka-python==2.0.2
    # via -r requirements.in
pydantic==1.8.2
    # via -r requirements.in
requests==2.26.0
    # via -r requirements.in
typer==0.4.0
    # via -r requirements.in
typing-extensions==3.10.0.2
    # via pydantic
urllib3==1.26.7
    # via requests
